+++ dflib/src/main/java/com/nhl/dflib/series/DoubleBaseSeries.java
@@ -4,12 +4,15 @@ import com.nhl.dflib.BooleanSeries;
 import com.nhl.dflib.DoubleSeries;
 import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.Series;
+import com.nhl.dflib.collection.BooleanMutableList;
 import com.nhl.dflib.collection.DoubleMutableList;
 import com.nhl.dflib.collection.IntMutableList;
 import com.nhl.dflib.collection.MutableList;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.filter.DoublePredicate;
 import com.nhl.dflib.filter.ValuePredicate;
+
+import java.util.Objects;
 
 import static java.util.Arrays.asList;
 
@@ -226,5 +229,56 @@ public abstract class DoubleBaseSeries implements DoubleSeries {
         }
 
         return vals.toSeries();
+    }
+
+    @Override
+    public BooleanSeries eq(Series<Double> another) {
+        int s = size();
+        int as = another.size();
+
+        if (s != as) {
+            throw new IllegalArgumentException("Another Series size " + as + " is not the same as this size " + s);
+        }
+
+        BooleanMutableList bools = new BooleanMutableList(s);
+
+        if (another instanceof DoubleSeries) {
+            DoubleSeries anotherDouble = (DoubleSeries) another;
+
+            for (int i = 0; i < s; i++) {
+                bools.add(getDouble(i) == anotherDouble.getDouble(i));
+            }
+        } else {
+            for (int i = 0; i < s; i++) {
+                bools.add(Objects.equals(get(i), another.get(i)));
+            }
+        }
+
+        return bools.toBooleanSeries();
+    }
+
+    @Override
+    public BooleanSeries ne(Series<Double> another) {
+        int s = size();
+        int as = another.size();
+
+        if (s != as) {
+            throw new IllegalArgumentException("Another Series size " + as + " is not the same as this size " + s);
+        }
+
+        BooleanMutableList bools = new BooleanMutableList(s);
+        if (another instanceof DoubleSeries) {
+            DoubleSeries anotherDouble = (DoubleSeries) another;
+
+            for (int i = 0; i < s; i++) {
+                bools.add(getDouble(i) != anotherDouble.getDouble(i));
+            }
+        } else {
+            for (int i = 0; i < s; i++) {
+                bools.add(!Objects.equals(get(i), another.get(i)));
+            }
+        }
+
+        return bools.toBooleanSeries();
     }
 }