+++ dflib/src/main/java/com/nhl/dflib/series/IntBaseSeries.java
@@ -3,11 +3,14 @@ package com.nhl.dflib.series;
 import com.nhl.dflib.BooleanSeries;
 import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.Series;
+import com.nhl.dflib.collection.BooleanMutableList;
 import com.nhl.dflib.collection.IntMutableList;
 import com.nhl.dflib.collection.MutableList;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.filter.IntPredicate;
 import com.nhl.dflib.filter.ValuePredicate;
+
+import java.util.Objects;
 
 import static java.util.Arrays.asList;
 
@@ -226,5 +229,56 @@ public abstract class IntBaseSeries implements IntSeries {
         }
 
         return vals.toSeries();
+    }
+
+    @Override
+    public BooleanSeries eq(Series<Integer> another) {
+        int s = size();
+        int as = another.size();
+
+        if (s != as) {
+            throw new IllegalArgumentException("Another Series size " + as + " is not the same as this size " + s);
+        }
+
+        BooleanMutableList bools = new BooleanMutableList(s);
+
+        if (another instanceof IntSeries) {
+            IntSeries anotherInt = (IntSeries) another;
+
+            for (int i = 0; i < s; i++) {
+                bools.add(getInt(i) == anotherInt.getInt(i));
+            }
+        } else {
+            for (int i = 0; i < s; i++) {
+                bools.add(Objects.equals(get(i), another.get(i)));
+            }
+        }
+
+        return bools.toBooleanSeries();
+    }
+
+    @Override
+    public BooleanSeries ne(Series<Integer> another) {
+        int s = size();
+        int as = another.size();
+
+        if (s != as) {
+            throw new IllegalArgumentException("Another Series size " + as + " is not the same as this size " + s);
+        }
+
+        BooleanMutableList bools = new BooleanMutableList(s);
+        if (another instanceof IntSeries) {
+            IntSeries anotherInt = (IntSeries) another;
+
+            for (int i = 0; i < s; i++) {
+                bools.add(getInt(i) != anotherInt.getInt(i));
+            }
+        } else {
+            for (int i = 0; i < s; i++) {
+                bools.add(!Objects.equals(get(i), another.get(i)));
+            }
+        }
+
+        return bools.toBooleanSeries();
     }
 }